Add CC0 license
Make the executive decision and license the project under CC0. Due to
the nature of the project, any similarity to licensed code is entirely
coincidental (but IANAL).

A project that is so widely touched by the public should be in the
public domain. It will also (probably) save us the trouble of having to
deal with licensing issues.